+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Release 1.5.1a7 (WIP)
+## Release 1.5.1a8 (WIP)
 
 ### Breaking Changes
 - Change default value for number of hyperparameter optimization trials from 10 to 500. (@ernestum)
@@ -14,6 +14,8 @@
 - Add `--max-total-trials` parameter to help with distributed optimization. (@ernestum)
 - Added `vec_env_wrapper` support in the config (works the same as `env_wrapper`)
 - Added Huggingface hub integration
+- Added `RecurrentPPO` support (aka `ppo_lstm`)
+- Added autodownload for "official" sb3 models from the hub
 
 ### Bug fixes
 - Fix `Reacher-v3` name in PPO hyperparameter file